Fix stride mode
